feat: move executor generic resolution to creation function
